# Conflicts:
#	Watchman.Discord/Areas/Users/Controllers/UsersController.cs
#	Watchman.DomainModel/DiscordServer/Commands/Handlers/SetRoleAsSafeCommandHandler.cs
#	Watchman.DomainModel/DiscordServer/Commands/Handlers/SetRoleAsUnsafeCommandHandler.cs
#	Watchman.DomainModel/DiscordServer/Queries/Handlers/GetDiscordServerSafeRolesQueryHandler.cs
#	Watchman.DomainModel/Help/HelpInformation.cs

+++ Watchman.Discord/Areas/Users/Controllers/UsersController.cs
@@ -1,4 +1,4 @@
-using Devscord.DiscordFramework.Commons.Extensions;
+using System.Collections.Generic;
 using Devscord.DiscordFramework.Framework.Architecture.Controllers;
 using Devscord.DiscordFramework.Framework.Commands.Responses;
 using Devscord.DiscordFramework.Middlewares.Contexts;
@@ -6,10 +6,11 @@ using Devscord.DiscordFramework.Services.Factories;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Devscord.DiscordFramework.Commons.Exceptions;
+using Devscord.DiscordFramework.Services;
 using Watchman.Cqrs;
 using Watchman.DomainModel.DiscordServer.Queries;
 using Watchman.Discord.Areas.Users.Services;
-using Devscord.DiscordFramework.Commons.Exceptions;
 using Watchman.Discord.Areas.Users.BotCommands;
 using Devscord.DiscordFramework.Services;
 
@@ -21,58 +22,84 @@ namespace Watchman.Discord.Areas.Users.Controllers
         private readonly MessagesServiceFactory _messagesServiceFactory;
         private readonly RolesService _rolesService;
         private readonly UsersService _usersService;
+        private readonly ResponsesService _responsesService;
 
-        public UsersController(IQueryBus queryBus, MessagesServiceFactory messagesServiceFactory, RolesService rolesService, UsersService usersService)
+        public UsersController(IQueryBus queryBus, MessagesServiceFactory messagesServiceFactory, RolesService rolesService, UsersService usersService, ResponsesService responsesService)
         {
             this._queryBus = queryBus;
             this._messagesServiceFactory = messagesServiceFactory;
             this._rolesService = rolesService;
             this._usersService = usersService;
+            this._responsesService = responsesService;
         }
 
-        public async Task GetAvatar(AvatarCommand avatarCommand, Contexts contexts)
+        public Task GetAvatar(AvatarCommand avatarCommand, Contexts contexts)
         {
             var messageService = this._messagesServiceFactory.Create(contexts);
             var user = contexts.User;
             if (avatarCommand.User != 0)
             {
-                user = await this._usersService.GetUserByIdAsync(contexts.Server, avatarCommand.User);
+                user = this._usersService.GetUserByIdAsync(contexts.Server, avatarCommand.User).GetAwaiter().GetResult();
             }
             if (string.IsNullOrEmpty(user.AvatarUrl))
             {
-                await messageService.SendResponse(x => x.UserDoesntHaveAvatar(user));
-                return;
+                return messageService.SendResponse(x => x.UserDoesntHaveAvatar(contexts.User));
             }
-            await messageService.SendMessage(user.AvatarUrl);
+            return messageService.SendMessage(contexts.User.AvatarUrl);
         }
 
-        public async Task AddRole(AddRoleCommand addRoleCommand, Contexts contexts)
+        public Task AddRole(AddRoleCommand addRoleCommand, Contexts contexts)
         {
+            if (addRoleCommand.Roles.Count > 5)
+            {
+                throw new InvalidArgumentsException();
+            }
             var safeRoles = this._queryBus.Execute(new GetDiscordServerSafeRolesQuery(contexts.Server.Id)).SafeRoles;
-            await this._rolesService.AddRoleToUser(safeRoles, contexts, addRoleCommand.Roles);
+            return this._rolesService.AddRoleToUser(safeRoles, contexts, addRoleCommand.Roles);
         }
 
-        public async Task RemoveRole(RemoveRoleCommand removeRoleCommand, Contexts contexts)
+        public Task RemoveRole(RemoveRoleCommand removeRoleCommand, Contexts contexts)
         {
+            if (removeRoleCommand.Roles.Count > 5)
+            {
+                throw new InvalidArgumentsException();
+            }
             var safeRoles = this._queryBus.Execute(new GetDiscordServerSafeRolesQuery(contexts.Server.Id)).SafeRoles;
-            await this._rolesService.DeleteRoleFromUser(safeRoles, contexts, removeRoleCommand.Roles);
+            return this._rolesService.DeleteRoleFromUser(safeRoles, contexts, removeRoleCommand.Roles);
         }
 
-        public async Task PrintRoles(RolesCommand rolesCommand, Contexts contexts)
+        public Task PrintRoles(RolesCommand rolesCommand, Contexts contexts)
         {
-            var messageService = this._messagesServiceFactory.Create(contexts);
+            var messagesService = this._messagesServiceFactory.Create(contexts);
             var query = new GetDiscordServerSafeRolesQuery(contexts.Server.Id);
             var safeRoles = this._queryBus.Execute(query).SafeRoles.ToList();
-
             if (!safeRoles.Any())
             {
-                await messageService.SendResponse(x => x.ServerDoesntHaveAnySafeRoles());
-                return;
+                return messagesService.SendResponse(x => x.ServerDoesntHaveAnySafeRoles());
             }
+            var serverRoles = contexts.Server.GetRoles();
+            var rolesNames = safeRoles
+                .Select(x => serverRoles.FirstOrDefault(serverRole => serverRole.Id == x.RoleId)?.Name)
+                .Where(x => x != default);
+            return this.SendRolesAsEmbedMessage(rolesNames, messagesService, contexts.Server.Id);
+        }
 
-            var output = new StringBuilder();
-            output.PrintManyLines(safeRoles.Select(x => x.Name).ToArray(), contentStyleBox: false, spacesBetweenLines: false);
-            await messageService.SendResponse(x => x.AvailableSafeRoles(output.ToString()));
+        private Task SendRolesAsEmbedMessage(IEnumerable<string> rolesNames, MessagesService messagesService, ulong serverId)
+        {
+            var title = this._responsesService.GetResponse(serverId, x => x.AvailableSafeRoles());
+            var description = this._responsesService.GetResponse(serverId, x => x.AvailableSafeRolesDescription());
+            var stringBuilder = new StringBuilder("**");
+            foreach (var role in rolesNames)
+            {
+                stringBuilder.AppendLine(role);
+            }
+            stringBuilder.Append("**");
+            var rolesResponse = this._responsesService.GetResponse(serverId, x => x.Roles()) + ":";
+            var values = new List<KeyValuePair<string, string>>
+            {
+                new KeyValuePair<string, string>(rolesResponse, stringBuilder.ToString())
+            };
+            return messagesService.SendEmbedMessage(title, description, values);
         }
     }
 }
